refactor(test): move check shape to operators tests.
