+++ tests/helpers/__init__.py
@@ -2,8 +2,6 @@
 
 from .asserts import assert_almost_allclose, assert_correlate
 from .factories import (
-    wrong_image_from_op,
-    wrong_kspace_from_op,
     kspace_from_op,
     image_from_op,
     to_interface,
@@ -16,8 +14,6 @@ from .factories import (
 __all__ = [
     "assert_almost_allclose",
     "assert_correlate",
-    "wrong_image_from_op",
-    "wrong_kspace_from_op",
     "kspace_from_op",
     "image_from_op",
     "to_interface",
